Factor out common code from translate header implementation
